// client/src/utils/types.ts
export type UserPeer = {
    id: string;
}



export type User = {
    id: number;
    username: string;
    firstName: string;
    lastName: string;
    profile?: Profile;
    presence?: UserPresence;
    peer: UserPeer;
};

export type UserPresence = {
    id: number;
    statusMessage?: string;
    showOffline: boolean;
}

export type Conversation = {
    id: number;
    creator: User;
    recipient: User;
    createdAt: string;
    lastMessageSent: MessageType;
};

export type CreateConversationParams = {
    username: string;
    message: string;
};

export type MessageType = {
    id: number;
    content: string;
    createdAt: string;
    author: User;
    conversation: Conversation;
};
// ...
export type EditMessagePayload = {
    id: number;
    messageId: number;
    content: string
}

export type ConversationType = "group" | "private"

export type ConversationTypeData = {
    type: ConversationType;
    label: string;
}

export type Group = {
    id: number;
    title?: String;
    users: User[];
    creator: User;
    owner: User;
    messages: GroupMessageType[];
    createdAt: number;
    lastMessageSent: MessageType;
    lastMessageSentAt: Date;
}

export type GroupMessageType = {
    id: number;
    content: string;
    createdAt: string;
    author: User;
    group: Group;


}

export type GroupMessage = {
    id: number;
    messages: GroupMessageType[];
};


export type FetchGroupMessagePayload = {
    id: number;
    messages: GroupMessageType[];
}

export type GroupMessageEventPayload = {
    message: GroupMessageType,
    group: Group;
}


export type CreateGroupParams = {
    users: string[];
    title: string;
}

export type DeleteGroupMessageResponse = {
    groupId: number;
    messageId: number;
}

export type DeleteGroupMessageParams = {
    id: number;
    messageId: number;
}
// ...
export type Profile = {
    id: number;
    about?: string;
    avatar?: string;
    banner?: string;
}